buildsys: refactor building and language initialization
rsync everything in a temporary build directory and let buildpackage
handle the rest.

Has the advantage that both packages can be assembled in one
buildpackage run, whereas previous it was called twice.
Further, we can omit the do `make install` manually before calling
dpkg-buildpackage and use .SECONDARY so that make keeps the files.

It's not used anywhere else in our build systems and not really
needed.

If a new language gets added just run:
make init-XY.po

This won't happen that often, so an auto create has not a real use,
especially as the .po file must then be edited anyway.
Also we now can reuse the new 'messages.pot' target in the 'update'
target and a strict separation between initialising a new language
and updating a existing one seems reasonable.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>