zh_CN: remove fuzzy from fixed lines
recently we got an update for our chinese translation with a lot of
fixes[1], but the external contributor forgot to remove the fuzzy
lines on fixed entries, so do this now. I left out commented out
entries for time reasons, we can always fix them up if they get
included again.

[1]: commit 59baa74279717cfcfb4b680f7bac9e970b93678cq