unify PO file license and copyright header
Add a SPDX header with the correct license to every file and use
year-less copyright, which is covered by the Berne Convention [0] and
other projects use this simpler approach nowadays too, e.g. curl [1].

For the contributors use a more general statement, as it's hard to
get a 100% complete list for each file, and that should be fine
w.r.t. CLA signed too. If somebody complains we can still check on
this more closely and maybe play Sherlock, but lets hope we can
invest our time better - for a more complete view see the git log.

[0]: https://en.wikipedia.org/wiki/Berne_Convention
[1]: https://daniel.haxx.se/blog/2023/01/08/copyright-without-years/

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>