update Spanish translations
We use quotation marks (“ ”) instead of double or single fullwidth
quotes (" ", ' ').

In principle the “Real Academia Española” (RAE) defines that angular
quotes (« ») is the preferable quotation and quotation marks as the next
option, however quotation marks are preferable in a more international
setting, e.g. the “Academia Mexicana de la Lengua” recommends to prefer
quotation marks due to its use in Mexico and the rest of America [1].

[1] https://es.wikipedia.org/wiki/Comillas