po2js: add ngettext support
We add new array for plural forms.

We use the msgid as the identifier translatable strings with a plural
form. This is the same as done for .mo files, from [1]:

> However, only the singular of the original string takes part in the
> hash table lookup.

We also obtain the plural forms and the number of plural forms directly
from the $LANG.po file, this requires the file to have a Plural-Forms
header that is valid, otherwise it would fail to evaluated by the
ngettext function when calling `Number($plural_forms)`.

Regarding the fallback to the original string, from [2]:

> The parameter n is used to determine the plural form. If no message
> catalog is found msgid1 is returned if n == 1, otherwise msgid2.

We default to the non-translated singular/plural depending on whether
n==1.

Regarding the default for $nplurals and $plural_forms: We select the
values for "Two forms, singular used for one only" from [2], this would
match with germanic and latin families.

This partially restores commit 2235ad7c7c79e29f09d3f3205ad86e862b5ac14c.

[1] https://www.gnu.org/software/gettext/manual/html_node/MO-Files.html
[2] https://www.gnu.org/software/gettext/manual/html_node/Plural-forms.html

Signed-off-by: Maximiliano Sandoval <m.sandoval@proxmox.com>
Link: https://lore.proxmox.com/20250731112822.297009-2-m.sandoval@proxmox.com
 [TL: fleece in follow-up fix from list reply]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>